requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
faiss-cpu==1.11.0
fastapi==0.116.0
filelock==3.18.0
fitz==0.0.1.dev2
fsspec==2025.5.1
hf-xet==1.1.5
httplib2==0.22.0
huggingface-hub==0.33.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.4
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
prov==2.1.1
puremagic==1.30
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.2.2
tqdm==4.67.1
traits==7.0.2
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
